changes to the scheduler
Each runnable task that we schedule is now stored in a map. the map
points each task to its corresponding thread. each thread is
responsible for checking if it is allowed to run a task, and then
sleeping itself until it can run again.

Beginning a new API that will allow passing messages to running tasks?