cleaned up more of the API
enhanced more of the API based around the Focus api.

now each API action for the scheduler builds upon a strategy that
atomically updates our task map and returns an IO action that can be
run to cancel, stop, rename tasks, etc.

next thing to do is remove dead code. lots of it.